move condition into own file so we can add tests
