+++ solr/core/src/java/org/apache/solr/update/processor/ConditionalUpsertProcessorFactory.java
@@ -16,27 +16,15 @@
  */
 package org.apache.solr.update.processor;
 
-import static org.apache.solr.common.SolrException.ErrorCode.SERVER_ERROR;
-import static org.apache.solr.update.processor.DistributingUpdateProcessorFactory.DISTRIB_UPDATE_PARAM;
-
 import java.io.IOException;
 import java.lang.invoke.MethodHandles;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
-import java.util.function.Function;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
-import org.apache.lucene.search.BooleanClause;
 import org.apache.lucene.util.BytesRef;
 import org.apache.solr.common.SolrException;
 import org.apache.solr.common.SolrInputDocument;
-import org.apache.solr.common.SolrInputField;
 import org.apache.solr.common.util.NamedList;
 import org.apache.solr.core.SolrCore;
 import org.apache.solr.handler.component.RealTimeGetComponent;
@@ -48,10 +36,13 @@ import org.apache.solr.util.plugin.SolrCoreAware;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import static org.apache.solr.common.SolrException.ErrorCode.SERVER_ERROR;
+import static org.apache.solr.update.processor.DistributingUpdateProcessorFactory.DISTRIB_UPDATE_PARAM;
+
 public class ConditionalUpsertProcessorFactory extends UpdateRequestProcessorFactory implements SolrCoreAware, UpdateRequestProcessorFactory.RunAlways {
   private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 
-  private final List<Condition> conditions = new ArrayList<>();
+  private final List<UpsertCondition> conditions = new ArrayList<>();
 
   @Override
   public void init(NamedList args)  {
@@ -60,7 +51,7 @@ public class ConditionalUpsertProcessorFactory extends UpdateRequestProcessorFac
       Object tmp = entry.getValue();
       if (tmp instanceof NamedList) {
         NamedList<String> condition = (NamedList<String>)tmp;
-        conditions.add(Condition.parse(name, condition));
+        conditions.add(UpsertCondition.parse(name, condition));
       }
       // TODO errors etc
     }
@@ -90,11 +81,11 @@ public class ConditionalUpsertProcessorFactory extends UpdateRequestProcessorFac
 
     private DistributedUpdateProcessor distribProc;  // the distributed update processor following us
     private DistributedUpdateProcessor.DistribPhase phase;
-    private final List<Condition> conditions;
+    private final List<UpsertCondition> conditions;
 
     ConditionalUpsertUpdateProcessor(SolrQueryRequest req,
                                      UpdateRequestProcessor next,
-                                     List<Condition> conditions) {
+                                     List<UpsertCondition> conditions) {
       super(next);
       this.core = req.getCore();
       this.conditions = conditions;
@@ -131,176 +122,20 @@ public class ConditionalUpsertProcessorFactory extends UpdateRequestProcessorFac
         SolrInputDocument newDoc = cmd.getSolrInputDocument();
         SolrInputDocument oldDoc = RealTimeGetComponent.getInputDocument(core, indexedDocId);
         if (oldDoc != null) {
-          Docs docs = new Docs(oldDoc, newDoc);
-          for (Condition condition: conditions) {
-            if (condition.matches(docs)) {
+          for (UpsertCondition condition: conditions) {
+            if (condition.matches(oldDoc, newDoc)) {
               log.info("Condition {} matched, taking action", condition.getName());
               if (condition.isSkip()) {
                 log.info("Condition {} matched - skipping insert", condition.getName());
                 return;
               }
-              condition.copyOldDocFields(docs);
+              condition.copyOldDocFields(oldDoc, newDoc);
               break;
             }
           }
         }
       }
       super.processAdd(cmd);
-    }
-  }
-
-  private static class Docs {
-    private final SolrInputDocument oldDoc;
-    private final SolrInputDocument newDoc;
-
-    Docs(SolrInputDocument oldDoc, SolrInputDocument newDoc) {
-      this.oldDoc = oldDoc;
-      this.newDoc = newDoc;
-    }
-
-    SolrInputDocument getOldDoc() {
-      return oldDoc;
-    }
-
-    SolrInputDocument getNewDoc() {
-      return newDoc;
-    }
-  }
-
-  private static class Condition {
-    private static final Pattern ACTION_PATTERN = Pattern.compile("^(skip)|(upsert):(\\*|[\\w,]+)$");
-    private static final List<String> ALL_FIELDS = Collections.singletonList("*");
-
-    private final String name;
-    private final List<FieldRule> rules;
-    private final boolean skip;
-    private final List<String> upsertFields;
-
-    Condition(String name, boolean skip, List<String> upsertFields, List<FieldRule> rules) {
-      this.name = name;
-      this.skip = skip;
-      this.upsertFields = upsertFields;
-      this.rules = rules;
-    }
-
-    static Condition parse(String name, NamedList<String> args) {
-      List<FieldRule> rules = new ArrayList<>();
-      boolean skip = false;
-      List<String> upsertFields = null;
-      for (Map.Entry<String, String> entry: args) {
-        String key = entry.getKey();
-        if ("action".equals(key)) {
-          String action = entry.getValue();
-          Matcher m = ACTION_PATTERN.matcher(action);
-          if (!m.matches()) {
-            throw new SolrException(SERVER_ERROR, "'" + action + "' not a valid action");
-          }
-          if (m.group(1) != null) {
-            skip = true;
-            upsertFields = null;
-          } else {
-            skip = false;
-            String fields = m.group(3);
-            upsertFields = Arrays.asList(fields.split(","));
-          }
-        } else {
-          BooleanClause.Occur occur = BooleanClause.Occur.valueOf(key.toUpperCase(Locale.ROOT));
-          String value = entry.getValue();
-          rules.add(FieldRule.parse(occur, value));
-        }
-      }
-      return new Condition(name, skip, upsertFields, rules);
-    }
-
-    String getName() {
-      return name;
-    }
-
-    boolean isSkip() {
-      return skip;
-    }
-
-    void copyOldDocFields(Docs docs) {
-      SolrInputDocument oldDoc = docs.getOldDoc();
-      SolrInputDocument newDoc = docs.getNewDoc();
-      Collection<String> fieldsToCopy;
-      if (ALL_FIELDS.equals(upsertFields)) {
-        fieldsToCopy = oldDoc.keySet();
-      } else {
-        fieldsToCopy = upsertFields;
-      }
-      fieldsToCopy.forEach(field -> {
-        if (!newDoc.containsKey(field)) {
-          SolrInputField inputField = oldDoc.getField(field);
-          newDoc.put(field, inputField);
-        }
-      });
-    }
-
-    boolean matches(Docs docs) {
-      boolean atLeastOneMatched = false;
-      for (FieldRule rule: rules) {
-        boolean ruleMatched = rule.matches(docs);
-        switch(rule.getOccur()) {
-          case MUST:
-            if (!ruleMatched) {
-              return false;
-            }
-            atLeastOneMatched = true;
-            break;
-          case MUST_NOT:
-            if (ruleMatched) {
-              return false;
-            }
-            break;
-          default:
-            atLeastOneMatched = ruleMatched || atLeastOneMatched;
-            break;
-        }
-      }
-      return atLeastOneMatched;
-    }
-  }
-
-  private static class FieldRule {
-    private static final Pattern RULE_CONDITION_PATTERN = Pattern.compile("^(OLD|NEW)\\.(\\w+):(\\w+)$");
-
-    private final BooleanClause.Occur occur;
-    private final Function<Docs, SolrInputDocument> docGetter;
-    private final String field;
-    private final String value;
-
-    private FieldRule(BooleanClause.Occur occur, Function<Docs, SolrInputDocument> docGetter, String field, String value) {
-      this.occur = occur;
-      this.docGetter = docGetter;
-      this.field = field;
-      this.value = value;
-    }
-
-    static FieldRule parse(BooleanClause.Occur occur, String condition) {
-      Matcher m = RULE_CONDITION_PATTERN.matcher(condition);
-      if (m.matches()) {
-        String doc = m.group(1);
-        String field = m.group(2);
-        String value = m.group(3);
-        Function<Docs, SolrInputDocument> docGetter;
-        if (doc.equalsIgnoreCase("OLD")) {
-          docGetter = Docs::getOldDoc;
-        } else {
-          docGetter = Docs::getNewDoc;
-        }
-        return new FieldRule(occur, docGetter, field, value);
-      }
-      throw new SolrException(SERVER_ERROR, "'" + condition + "' not a valid condition for rule");
-    }
-
-    BooleanClause.Occur getOccur() {
-      return occur;
-    }
-
-    boolean matches(Docs docs) {
-      SolrInputDocument doc = docGetter.apply(docs);
-      return value.equals(doc.getFieldValue(field));
     }
   }
 }
